pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cdc-pipeline-generator"
version = "0.10.2"
description = "CDC pipeline configuration generator for Redpanda Connect"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Igor Efrem", email = "igor@carasent.com"}
]
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pyyaml>=6.0",
    "ruamel.yaml>=0.17",
    "jinja2>=3.1",
]

[project.optional-dependencies]
# Database connectivity - heavy dependencies for DB inspection
database = [
    "pymssql>=2.2",
    "psycopg2-binary>=2.9",
]

# Full installation - everything needed for development/production
full = [
    "pymssql>=2.2",
    "psycopg2-binary>=2.9",
]

# Development tools
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/carasent/cdc-pipeline-generator"
Repository = "https://github.com/carasent/cdc-pipeline-generator"
"Bug Tracker" = "https://github.com/carasent/cdc-pipeline-generator/issues"

[project.scripts]
cdc = "cdc_generator.cli.commands:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["cdc_generator*"]

[tool.setuptools.package-data]
cdc_generator = [
    "templates/init/*",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "ANN",    # flake8-annotations
    "ASYNC",  # flake8-async
    "B",      # flake8-bugbear
    "A",      # flake8-builtins
    "C4",     # flake8-comprehensions
    "RET",    # flake8-return
    "SIM",    # flake8-simplify
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
    "PL",     # pylint
    "RUF",    # ruff-specific rules
]

[tool.ruff.lint.per-file-ignores]
# Allow Any in type stub files and tests if needed
"*.pyi" = ["ANN401"]
"tests/*" = ["ANN401", "ARG"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_explicit = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_unreachable = true
